fix: move mails into src
