+++ src/Loyalty/Mail/AccountActivated.php
@@ -1,27 +1,25 @@
 <?php
 
-namespace App\Mail;
+namespace Src\Loyalty\Mail;
 
 use Illuminate\Bus\Queueable;
 use Illuminate\Mail\Mailable;
 use Illuminate\Queue\SerializesModels;
 
-class LoyaltyPointsReceived extends Mailable
+class AccountActivated extends Mailable
 {
     use Queueable, SerializesModels;
 
     private $balance;
-    private $pointsAmount;
 
     /**
      * Create a new message instance.
      *
      * @return void
      */
-    public function __construct($pointsAmount, $balance)
+    public function __construct($balance)
     {
         $this->balance = $balance;
-        $this->pointsAmount = $pointsAmount;
     }
 
     /**
@@ -31,9 +29,8 @@ class LoyaltyPointsReceived extends Mailable
      */
     public function build()
     {
-        return $this->view('emails.loyaltyPointsReceived')->with([
+        return $this->view('emails.accountActivated')->with([
             'balance' => $this->balance,
-            'points' => $this->pointsAmount,
         ]);
     }
 }
