+++ src/Loyalty/Mail/AccountDeactivated.php
@@ -1,9 +1,8 @@
 <?php
 
-namespace App\Mail;
+namespace Src\Loyalty\Mail;
 
 use Illuminate\Bus\Queueable;
-use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Mail\Mailable;
 use Illuminate\Queue\SerializesModels;
